Removed init functions
