+++ cmd/root.go
@@ -2,17 +2,17 @@ package cmd
 
 import (
 	"fmt"
+	"github.com/elahe-dastan/urlShortener_KGS/cmd/server"
+	"github.com/elahe-dastan/urlShortener_KGS/cmd/setupdb"
 	"github.com/elahe-dastan/urlShortener_KGS/config"
 	"github.com/elahe-dastan/urlShortener_KGS/db"
 	"github.com/elahe-dastan/urlShortener_KGS/middleware"
 	"github.com/spf13/cobra"
 	"os"
-
-	homedir "github.com/mitchellh/go-homedir"
-	"github.com/spf13/viper"
 )
 
 var cfgFile string
+var exitFailure = 1
 
 // rootCmd represents the base command when called without any subcommands
 var rootCmd = &cobra.Command {
@@ -24,52 +24,14 @@ var rootCmd = &cobra.Command {
 // Execute adds all child commands to the root command and sets flags appropriately.
 // This is called by main.main(). It only needs to happen once to the rootCmd.
 func Execute() {
-	if err := rootCmd.Execute(); err != nil {
-		fmt.Println(err)
-		os.Exit(1)
-	}
-}
-
-func init() {
-	cobra.OnInitialize(initConfig)
-
-	// Here you will define your flags and configuration settings.
-	// Cobra supports persistent flags, which, if defined here,
-	// will be global for your application.
-
-	rootCmd.PersistentFlags().StringVar(&cfgFile, "config", "", "config file (default is $HOME/.urlShortener.yaml)")
-
-	// Cobra also supports local flags, which will only run
-	// when this action is called directly.
-	rootCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
-}
-
-// initConfig reads in config file and ENV variables if set.
-func initConfig() {
-	if cfgFile != "" {
-		// Use config file from the flag.
-		viper.SetConfigFile(cfgFile)
-	} else {
-		// Find home directory.
-		home, err := homedir.Dir()
-		if err != nil {
-			fmt.Println(err)
-			os.Exit(1)
-		}
-
-		// Search config in home directory with name ".urlShortener" (without extension).
-		viper.AddConfigPath(home + "/go/src/urlShortener/config")
-		viper.SetConfigName("config")
-		viper.SetConfigType("yml")
-	}
-
-	viper.AutomaticEnv() // read in environment variables that match
-
 	constant := config.ReadConfig()
 	db.SetConfig(constant)
 	middleware.SetConfig(constant)
-	//// If a config file is found, read it in.
-	//if err := viper.ReadInConfig(); err == nil {
-	//	fmt.Println("Using config file:", viper.ConfigFileUsed())
-	//}
+
+	setupdb.Register(rootCmd)
+	server.Register(rootCmd)
+	if err := rootCmd.Execute(); err != nil {
+		fmt.Println(err)
+		os.Exit(exitFailure)
+	}
 }
