--- cmd/setupdb.go
@@ -1,48 +1,0 @@
-/*
-Copyright © 2020 NAME HERE <EMAIL ADDRESS>
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-package cmd
-
-import (
-	"github.com/elahe-dastan/urlShortener_KGS/db"
-	"github.com/spf13/cobra"
-)
-
-// setupdbCmd represents the setupdb command
-var setupdbCmd = &cobra.Command{
-	Use:   "setupdb",
-	Short: "Manages database, creates and fills tables if don't exist",
-	Run: func(cmd *cobra.Command, args []string) {
-		//This part of code generates all the random short ULRs possible and saves them into the database
-		//This code executes only if the table containing the short URLs doesn't exist
-		db.SaveShortURLs()
-		//Creates a table for mapping each long URL to a short URL if not present in the database
-		db.CreateMap()
-	},
-}
-
-func init() {
-	rootCmd.AddCommand(setupdbCmd)
-
-	// Here you will define your flags and configuration settings.
-
-	// Cobra supports Persistent Flags which will work for this command
-	// and all subcommands, e.g.:
-	// setupdbCmd.PersistentFlags().String("foo", "", "A help for foo")
-
-	// Cobra supports local flags which will only run when this command
-	// is called directly, e.g.:
-	// setupdbCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
-}
